pyproject.toml
# ==============================
# AI Underwriting Project Configuration
# ==============================

[project]
name = "ai-underwriting"
version = "0.1.0"
description = "AI-powered underwriting system"
authors = [{name = "Kunal Sharma", email = "kunal.ucet@gmail.com"}]
readme = "README.md"
requires-python = "=3.10.13"
license = "MIT"

# ------------------------------
# Dependencies
# ------------------------------
dependencies = [
    # Web Framework & API Development
    "Flask>=3.1.0",
    "Flask-Cors>=5.0.1",
    "Flask-RESTful>=0.3.9",
    "gunicorn>=20.1.0",

    # LangChain (AI Agent Framework)
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0,<1.0",
    "langsmith>=0.2.0",
    "langchain-openai>=0.3.8",

    # OpenAI Integration
    "openai>=1.3.0",

    # Database & ORM
    "SQLAlchemy>=2.0.20",
    "Flask-SQLAlchemy>=3.0.5",
    "psycopg2-binary>=2.9.7",
    "pymysql>=1.1.0",

    # Machine Learning & Data Processing
    "numpy>=1.24.4",
    "pandas>=2.0.3",
    "scikit-learn>=1.2.2",

    # API & Web Requests Handling
    "requests>=2.31.0",

    # Logging & Monitoring
    "loguru>=0.7.0",

    # Environment Variable Management
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",

    # Redis & Background Task Processing
    "redis>=5.0.0",
    "rq>=1.15.0"
]

# ------------------------------
# Poetry Configuration
# ------------------------------
[tool.poetry]
packages = [{include = "src"}]  # Ensure Poetry knows where your package code is

# ------------------------------
# Development Dependencies
# ------------------------------
[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^4.0"
faker = "^19.6.0"
locust = "^2.20.0"
flask-limiter = "^3.5.0"
pytest-html = "^4.0.0"
pytest-selenium = "^4.0.0"
pytest-redis = "^3.0.0"
pytest-asyncio = "^0.23.6"

# ------------------------------
# Build System Configuration
# ------------------------------
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
